pyproject.toml
[tool.poetry]
name = "t4-cicd"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "LICENSE"
readme = "README.md"

# For new package, add it here
packages = [
    {include = "cli", from="src"},
    {include = "controller", from="src"},
    {include = "util", from="src"},
]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
gitpython = "^3.1"
pyyaml = "^6.0" 
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
pathlib = "^1.0.1"
ruamel-yaml = "^0.18.6"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
mongomock = "^4.2.0.post1"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report html --cov-fail-under=50"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cid = "cli.__main__:cid"